Moving the code for Ai to seprate file
